+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Add `Shaped::Array` method and class
 - Add `Array` refinement
+- Add `Shaped::MatchFailureReason` and `Shaped::Array#match_failure_reason`
 
 ## 0.1.0 - 2020-06-16
 - Initial release of Shaped! Validate the shape of Ruby hashes and/or arrays.